elasticsearch集成以及文档完善

+++ akkad-springboot/springboot-dist-trans/dist-trans-atomikos/src/main/java/xyz/wongs/weathertop/mapper/location/LocationMapper.java
@@ -1,8 +1,8 @@
 package xyz.wongs.weathertop.mapper.location;
 
 
-import xyz.wongs.weathertop.base.persistence.mybatis.annotation.MyBatisMapper;
-import xyz.wongs.weathertop.base.persistence.mybatis.mapper.BaseMapper;
+import xyz.wongs.drunkard.base.persistence.mybatis.annotation.MyBatisMapper;
+import xyz.wongs.drunkard.base.persistence.mybatis.mapper.BaseMapper;
 import xyz.wongs.weathertop.entity.location.Location;
 
 import java.util.List;
@@ -11,7 +11,7 @@ import java.util.List;
 public interface LocationMapper extends BaseMapper<Location,Long> {
     int deleteByPrimaryKey(Long id);
 
-    Long insertSelective(Location record);
+    int insertSelective(Location record);
 
     Location selectByPrimaryKey(Long id);
